feat: add syncer interface for defining upload/download data between cloud storage and local disk
